server/requirements.txt
coloredlogs
Flask
psycopg2-binary
flask-socketio
eventlet
dataclasses
cachetools
Flask-BasicAuth
flask-cors
